langgraph-digital-marketer/pyproject.toml
[project]
name = "langgraph-digital-marketer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=2.0.9",
    "langgraph>=0.2.74",
    "python-dotenv>=1.0.1",
]
